Move capybara config to separate file
